send notification when rule does not match
In my case a lot of regional product have short codes that match in the UPC
database as something completely different to what they really are. This results
in lot of "don't know what to do with..." messages that are visible only in the
log files. I think it's much better if user gets at least notified about them.